move get-babel-options out of webpack config
