+++ get-webpack-config.js
@@ -3,6 +3,7 @@
 const path = require('path');
 const webpack = require('webpack');
 const { isDebug, buildDir } = require('./build-arguments');
+const getBabelOptions = require('./get-babel-options');
 
 const HappyPack = require('happypack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
@@ -25,7 +26,6 @@ module.exports = function getWebpackConfig({
     getLoaderForModule,
     tsConfigPath,
 }) {
-    console.log('123');
     return {
         mode: 'development',
         devtool: 'cheap-module-eval-source-map',
@@ -179,80 +179,3 @@ module.exports = function getWebpackConfig({
         ],
     };
 };
-
-function getBabelOptions({ isReactNative, getBabelConfig }) {
-    let babelOverrides = {
-        compact: false,
-        minified: false,
-        cacheDirectory: true,
-    };
-
-    if (getBabelConfig) {
-        const babelConfig = getBabelConfig({ path });
-
-        if (!babelConfig.plugins) {
-            babelConfig.plugins = [];
-        }
-
-        babelConfig.plugins.unshift(require.resolve('react-hot-loader/babel'));
-
-        return Object.assign({}, babelConfig, babelOverrides);
-    }
-
-    // @TODO - rethink this, should clients always pass their babel config? Or should we auto-detect it?
-    if (isReactNative) {
-        return Object.assign(
-            {
-                babelrc: false,
-                presets: [require.resolve('metro-react-native-babel-preset')],
-                plugins: [require.resolve('react-hot-loader/babel')],
-            },
-            babelOverrides
-        );
-    }
-
-    // @TODO - rethink this, should clients always pass their babel config? Or should we auto-detect it?
-    return Object.assign(
-        {
-            babelrc: false,
-            presets: [
-                [
-                    require.resolve('@babel/preset-env'),
-                    {
-                        targets: {
-                            ie: 11,
-                        },
-                    },
-                ],
-                [
-                    require.resolve('@babel/preset-react'),
-                    {
-                        development: true,
-                    },
-                ],
-                [
-                    '@babel/preset-typescript',
-                    {
-                        isTSX: true,
-                        allExtensions: true,
-                    },
-                ],
-            ],
-            plugins: [
-                require.resolve('react-hot-loader/babel'),
-
-                // Needed to have parity with TS class properties
-                '@babel/plugin-proposal-class-properties',
-
-                // Allow experimental rest spread syntax
-                [
-                    '@babel/plugin-proposal-object-rest-spread',
-                    {
-                        useBuiltIns: true,
-                    },
-                ],
-            ],
-        },
-        babelOverrides
-    );
-}
